backend/requirements.txt
annotated-types
anyio
cachetools
certifi
charset-normalizer
click
dnspython
docstring_parser
email-validator
fastapi
fastapi-cli
fastapi-cloud-cli
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-genai
google-generativeai
google-resumable-media
googleapis-common-protos
grpc-google-iam-v1
grpcio
grpcio-status
gunicorn
h11
httpcore
httplib2
httptools
httpx
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
mutagen
nest-asyncio
numpy
packaging
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
python-multipart
python-slugify
PyYAML
requests
rich
rich-toolkit
rignore
rsa
sentry-sdk
shapely
shellingham
six
sniffio
starlette
tenacity
text-unidecode
tqdm
typer
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
uvloop
watchfiles
websockets
